Client/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64. ubuntu 22.04
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
backcall=0.2.0=pypi_0
blas=1.0=openblas
boto3=1.26.57=pypi_0
botocore=1.29.57=pypi_0
bzip2=1.0.8=h7f98852_4
ca-certificates=2022.12.7=ha878542_0
cachetools=5.3.0=pypi_0
cairo=1.16.0=h18b612c_1001
certifi=2022.12.7=pyhd8ed1ab_0
charset-normalizer=3.0.1=pypi_0
click=8.0.4=pypi_0
cycler=0.11.0=pyhd8ed1ab_0
dbus=1.13.18=hb2f20db_0
decorator=5.1.1=pypi_0
eigen=3.4.0=h4bd325d_0
expat=2.2.10=h9c3ff4c_0
ffmpeg=4.2.2=h20bf706_0
ffmpeg-python=0.2.0=py_0
filelock=3.9.0=pypi_0
fire=0.5.0=pypi_0
fontconfig=2.14.1=hef1e5e3_0
fonttools=4.38.0=pypi_0
freetype=2.10.4=h0708190_1
future=0.18.2=py37h89c1867_5
giflib=5.2.1=h36c2ea0_2
gitdb=4.0.10=pypi_0
gitpython=3.1.30=pypi_0
glib=2.69.1=he621ea3_2
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
google-auth=2.16.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
graphite2=1.3.14=h295c915_1
grpcio=1.51.1=pypi_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
harfbuzz=4.3.0=hd55b92a_0
hdf5=1.10.6=h3ffc7dd_1
huggingface-hub=0.12.0=pypi_0
icu=58.2=hf484d3e_1000
idna=3.4=pypi_0
importlib-metadata=6.0.0=pypi_0
imutils=0.5.4=py37h89c1867_2
ipython=7.34.0=pypi_0
jedi=0.18.2=pypi_0
jmespath=1.0.1=pypi_0
jpeg=9e=h166bdaf_1
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.4=py37h6a678d5_0
krb5=1.19.3=h3790be6_0
lame=3.100=h7f98852_1001
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libblas=3.9.0=15_linux64_openblas
libcblas=3.9.0=15_linux64_openblas
libclang=10.0.1=default_hb85057a_2
libdeflate=1.8=h7f8727e_5
libedit=3.1.20191231=he28a2e2_2
libevent=2.1.12=h8f2d780_0
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=12.2.0=h69a702a_19
libgfortran5=12.2.0=h337968e_19
libgomp=11.2.0=h1234567_1
liblapack=3.9.0=15_linux64_openblas
libllvm10=10.0.1=he513fc3_3
libopenblas=0.3.20=pthreads_h78a6416_0
libopus=1.3.1=h7f98852_1
libpng=1.6.37=hbc83047_0
libpq=12.9=h16c4e8d_3
libprotobuf=3.20.1=h4ff587b_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.4.0=hecacb30_2
libuuid=2.32.1=h7f98852_1000
libvpx=1.7.0=h439df22_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.3=he3ba5ed_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
lz4-c=1.9.3=h9c3ff4c_1
markdown=3.4.1=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.5.3=pypi_0
matplotlib-inline=0.1.6=pypi_0
ncurses=6.3=h5eee18b_3
nettle=3.6=he412f7d_0
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numpy=1.21.6=py37h976b520_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
oauthlib=3.2.2=pypi_0
openblas=0.3.20=pthreads_h320a7e8_0
opencv=4.6.0=py37hd653453_2
opencv-python=4.7.0.68=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.3.1=hf7af979_3
openssl=1.1.1s=h7f8727e_0
packaging=23.0=pypi_0
pandas=1.3.5=pypi_0
parso=0.8.3=pypi_0
pcre=8.45=h9c3ff4c_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.4.0=pypi_0
pip=22.3.1=py37h06a4308_0
pixman=0.38.0=h516909a_1003
prompt-toolkit=3.0.36=pypi_0
protobuf=3.20.3=pypi_0
psutil=5.9.4=pypi_0
ptyprocess=0.7.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pybboxes=0.1.6=pypi_0
pygments=2.14.0=pypi_0
pyparsing=3.0.9=pyhd8ed1ab_0
python=3.7.16=h7a1cb2a_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.7=2_cp37m
pytz=2022.7.1=pyhd8ed1ab_0
pyyaml=6.0=pypi_0
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtwebkit=5.212=h4eab89a_4
readline=8.2=h5eee18b_0
requests=2.28.2=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
s3transfer=0.6.0=pypi_0
sahi=0.11.11=pypi_0
scipy=1.7.3=py37hf2a6cf1_0
seaborn=0.12.2=pypi_0
setuptools=65.6.3=py37h06a4308_0
shapely=2.0.0=pypi_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.0=pypi_0
sqlite=3.40.1=h5082296_0
tensorboard=2.11.2=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
termcolor=2.2.0=pypi_0
terminaltables=3.1.10=pypi_0
thop=0.1.1-2209072238=pypi_0
tk=8.6.12=h1ccaba5_0
torch=1.13.1=pypi_0
torchvision=0.14.1=pypi_0
tornado=6.1=py37h540881e_3
tqdm=4.64.1=pypi_0
traitlets=5.8.1=pypi_0
typing_extensions=4.4.0=pyha770c72_0
urllib3=1.26.14=pypi_0
wcwidth=0.2.6=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
x264=1!157.20191217=h7b6447c_0
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.0.10=h7f98852_0
xorg-libsm=1.2.3=hd9c2040_1000
xorg-libx11=1.7.2=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.10=h5eee18b_1
yolov5=7.0.7=pypi_0
yolov7detect=1.0.0=pypi_0
zipp=3.11.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
